+++ README.md
@@ -1,9 +1,6 @@
 # Using-machine-learning-to-detect-maclicious-URLs
 
-With the growth of Machine Learning in the past few years, many tasks are being done with the help of machine learning algorithms.Unfortunately or fortunately, there has been little work done on security with machine learning algorithms. So I thought of presenting some at Fsecurify.
-
-Using machine learning to detect maclicious URLs, We at Fsecurify used machine learning to predict whether a URL is good or bad/malicious/phishing. The results are great and we get about 98% accuracy. The code and the tutorial is given on the website.
-
+``` This Repo is based on https://github.com/faizann24/Using-machine-learning-to-detect-malicious-URLs ```
 http://fsecurify.com/using-machine-learning-detect-malicious-urls/
