+++ data/.~lock.data.csv#
@@ -1,0 +1,1 @@
+Joerg Stephan,johest,localhost.localdomain,18.02.2017 11:31,file:///home/johest/.config/libreoffice/4;
